Move tests back to src directory
