+++ src/index.test.js
@@ -1,4 +1,4 @@
-import SkynetClient from "../src/index";
+import SkynetClient from "./index";
 
 describe("SkynetClient", () => {
   it("should contain all api methods", () => {
